Introduce pyfakefs for thumbnail tests
