+++ src/requirements.txt
@@ -15,4 +15,6 @@ requests==2.25.1
 # Default is master branch specified in docker-compose.yml if not set
 git+git://github.com/hubmapconsortium/commons.git@${COMMONS_BRANCH}#egg=hubmap-commons
 
+# For unit test
 nose2==0.10.0
+pyfakefs==4.5.0
